Added memory management
