--- Asak/src/kernel/memory/mem.cpp
@@ -1,19 +1,0 @@
-// Overload the new operator for single object allocation
-void* operator new(std::size_t size) {
-    return malloc(size);   // Call the C standard library function malloc() to allocate memory of the given size and return a pointer to it
-}
-
-// Overload the delete operator for single object deallocation
-void operator delete(void* ptr) noexcept {
-    free(ptr);             // Call the C standard library function free() to deallocate the memory pointed to by the given pointer
-}
-
-// Overload the new operator for array allocation
-void* operator new[](std::size_t size) {
-    return malloc(size);   // Call the C standard library function malloc() to allocate memory of the given size and return a pointer to it
-}
-
-// Overload the delete operator for array deallocation
-void operator delete[](void* ptr) noexcept {
-    free(ptr);             // Call the C standard library function free() to deallocate the memory pointed to by the given pointer
-}
